+++ crates/chain/src/persist.rs
@@ -1,26 +1,26 @@
-use core::marker::PhantomData;
-
 use crate::Append;
 
 /// `Persist` wraps a [`PersistBackend`] (`B`) to create a convenient staging area for changes (`C`)
 /// before they are persisted.
 ///
-/// Not all changes to the tracker (`T`), which is an in-memory representation of wallet/blockchain
+/// Not all changes to the tracker, which is an in-memory representation of wallet/blockchain
 /// data, needs to be written to disk right away, so [`Persist::stage`] can be used to *stage*
 /// changes first and then [`Persist::commit`] can be used to write changes to disk.
-pub struct Persist<T, B, C> {
+pub struct Persist<B, C> {
     backend: B,
     stage: C,
-    tracker: PhantomData<T>,
 }
 
-impl<T, B: PersistBackend<T, C>, C: Append + Default> Persist<T, B, C> {
+impl<B, C> Persist<B, C>
+where
+    B: PersistBackend<C>,
+    C: Append + Default,
+{
     /// Create a new [`Persist`] from [`PersistBackend`].
     pub fn new(backend: B) -> Self {
         Self {
             backend,
             stage: Default::default(),
-            tracker: Default::default(),
         }
     }
 
@@ -42,20 +42,16 @@ impl<T, B: PersistBackend<T, C>, C: Append + Default> Persist<T, B, C> {
     pub fn commit(&mut self) -> Result<(), B::WriteError> {
         let mut temp = C::default();
         core::mem::swap(&mut temp, &mut self.stage);
-        self.backend.write_changes(temp)
+        self.backend.write_changes(&temp)
     }
 }
 
 /// A persistence backend for [`Persist`].
 ///
-/// * `T` represents the tracker.
-/// * `C` represents the changeset.
-pub trait PersistBackend<T, C> {
+/// `C` represents the changeset.
+pub trait PersistBackend<C> {
     /// The error the backend returns when it fails to write.
     type WriteError: core::fmt::Debug;
-
-    /// The error the backend returns when it fails to load.
-    type LoadError: core::fmt::Debug;
 
     /// Writes a changeset to the persistence backend.
     ///
@@ -64,20 +60,29 @@ pub trait PersistBackend<T, C> {
     /// that [`load_into_tracker`] restores a keychain tracker to what it should be if all
     /// changesets had been applied sequentially.
     ///
-    /// [`load_into_tracker`]: Self::load_into_tracker
-    fn write_changes(&mut self, changeset: C) -> Result<(), Self::WriteError>;
+    /// [`load_into_tracker`]: LoadablePersistBackend::load_into_tracker
+    fn write_changes(&mut self, changeset: &C) -> Result<(), Self::WriteError>;
+}
+
+/// A trait that extends [`PersistBackend`] to be able to load into a tracker implementation (`T`).
+pub trait LoadablePersistBackend<T, C>: PersistBackend<C> {
+    /// The error the backend returns when it fails to load.
+    type LoadError: core::fmt::Debug;
 
     /// Loads all data from the persistence backend into `tracker`.
     fn load_into_tracker(&mut self, tracker: &mut T) -> Result<(), Self::LoadError>;
 }
 
-impl<T, C> PersistBackend<T, C> for () {
+impl<C> PersistBackend<C> for () {
     type WriteError = ();
-    type LoadError = ();
 
-    fn write_changes(&mut self, _changeset: C) -> Result<(), Self::WriteError> {
+    fn write_changes(&mut self, _changeset: &C) -> Result<(), Self::WriteError> {
         Ok(())
     }
+}
+
+impl<T, C> LoadablePersistBackend<T, C> for () {
+    type LoadError = ();
 
     fn load_into_tracker(&mut self, _tracker: &mut T) -> Result<(), Self::LoadError> {
         Ok(())
