+++ crates/file_store/src/lib.rs
@@ -4,10 +4,6 @@ mod keychain_store;
 mod store;
 use std::io;
 
-use bdk_chain::{
-    keychain::{KeychainChangeSet, KeychainTracker, PersistBackend},
-    sparse_chain::ChainPosition,
-};
 use bincode::{DefaultOptions, Options};
 pub use entry_iter::*;
 pub use keychain_store::*;
@@ -15,31 +11,6 @@ pub use store::*;
 
 pub(crate) fn bincode_options() -> impl bincode::Options {
     DefaultOptions::new().with_varint_encoding()
-}
-
-impl<K, P> PersistBackend<K, P> for KeychainStore<K, P>
-where
-    K: Ord + Clone + core::fmt::Debug,
-    P: ChainPosition,
-    KeychainChangeSet<K, P>: serde::Serialize + serde::de::DeserializeOwned,
-{
-    type WriteError = std::io::Error;
-
-    type LoadError = IterError;
-
-    fn append_changeset(
-        &mut self,
-        changeset: &KeychainChangeSet<K, P>,
-    ) -> Result<(), Self::WriteError> {
-        KeychainStore::append_changeset(self, changeset)
-    }
-
-    fn load_into_keychain_tracker(
-        &mut self,
-        tracker: &mut KeychainTracker<K, P>,
-    ) -> Result<(), Self::LoadError> {
-        KeychainStore::load_into_keychain_tracker(self, tracker)
-    }
 }
 
 /// Error that occurs due to problems encountered with the file.
